refactor(ui): split settings.jsx into SettingsShell + ESM pages (P3-ui MVP)
Splits the 2598-line window-globals settings.jsx monolith into a real ES
module tree per spec-settings.md (c): SettingsShell (nav + section
routing) + SettingsNav (grouped SERVER/MODELS/INFERENCE/ROUTING/
OBSERVABILITY/DATA/DIAGNOSTICS/INTEGRATIONS rail) + one page module per
section under settings/pages/<group>/, plus settings/shared/ for
ApplyBadge, the schema-driven row engine (AdvRow/_schemaField/_advCoerce/
_deepMergePatch/_getIn), SRow, and RestartApiPanel.

All 12 existing sections (general/slots/npu/memory/agents/voice/imagegen/
storage/secrets/updates/advanced/about) move verbatim with unchanged
`id`s, so #settings/<id> deep links keep working; only labels/grouping
changed. Threaded the implicit window-globals the old file relied on
(Icons, ConfirmDialog, AddSecretModal, React hooks) as real imports —
added `export` to their existing declarations in chrome.jsx/primitives.jsx/
extra-modals.jsx without touching their window-globals publish, which
other not-yet-migrated dash/*.jsx files still depend on. window.__hal0Toast
stays a global (runtime-installed toast singleton, not a static export).

Adds four new nav slots per the MVP cut: Security and Hardware Tuning are
visible-but-disabled (blocked on the auth/tuning lanes); Library &
Downloads and Health & Stats are visible stub pages (their backend routes
exist but have no settings UI yet) rather than fabricated new features.
Doctor is a minimal stub (CLI-only today, no HTTP endpoint). Backend&GPU
and Model Defaults are left as TODO(session3 ML-4) markers in
SettingsNav.jsx, not built.

main.jsx now imports SettingsShell directly instead of reading
window.SettingsView; removed the settings.jsx side-effect import from
main.tsx and deleted the old file.

ui/dash/build + ui/dash/lint pass clean.
